Moved spmdm to main repo
